refactor(client): organize api calls into single folder
Refactors the client-side code by renaming services directories to api and updating all relevant
imports. This change improves code organization by centralizing all API-related functions into a
single, clearly-named directory within each feature module.
