+++ apps/client/src/features/auth/api/index.ts
@@ -1,0 +1,8 @@
+// Local Auth Api
+export * from "./auth-api";
+
+// Password Reset Api
+export * from "./reset-password-api";
+
+// Verify Email Api
+export * from "./verify-email-api";
