+++ apps/client/src/features/auth/api/verify-email-api.ts
