+++ apps/client/src/features/auth/queries/use-reset-password-mutation.ts
@@ -4,7 +4,7 @@ import { useMutation } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import toast from "react-hot-toast";
 
-import { resetPassword } from "../services";
+import { resetPassword } from "../api";
 
 export function useResetPasswordMutation() {
   const router = useRouter();
