+++ apps/client/src/features/auth/queries/use-verify-password-reset-token-query.tsx
@@ -2,7 +2,7 @@
 
 import { useQuery } from "@tanstack/react-query";
 
-import { verifyPasswordResetToken } from "../services";
+import { verifyPasswordResetToken } from "../api";
 
 export function useVerifyPasswordResetTokenQuery(token: string) {
   return useQuery({
