--- apps/client/src/features/auth/services/index.ts
@@ -1,8 +1,0 @@
-// Local Auth Services
-export * from "./auth.service";
-
-// Password Reset Services
-export * from "./reset-password.service";
-
-// Verify Email Service
-export * from "./verify-email.service";
