+++ apps/client/src/features/note/queries/use-get-notes-count-by-video-id-query.ts
@@ -2,7 +2,7 @@
 
 import { useQuery } from "@tanstack/react-query";
 
-import { getNotesCountByVideoId } from "../services";
+import { getNotesCountByVideoId } from "../api";
 
 export function useGetNotesCountByVideoIdQuery(ytVideoId: string) {
   return useQuery({
