--- apps/client/src/features/note/services/index.ts
@@ -1,1 +1,0 @@
-export * from "./note.service";
