+++ apps/client/src/features/user/queries/use-update-user-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 
-import { updateCurrentUser } from "../services";
+import { updateCurrentUser } from "../api";
 
 export function useUpdateUserMutation() {
   const queryClient = useQueryClient();
