--- apps/client/src/features/user/services/index.ts
@@ -1,1 +1,0 @@
-export * from "./user.service";
