+++ apps/client/src/features/video/api/index.ts
@@ -1,0 +1,1 @@
+export * from "./video-api";
