+++ apps/client/src/features/video/hooks/use-save-video-data.tsx
@@ -4,7 +4,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRouter } from "next/navigation";
 import toast from "react-hot-toast";
 
-import { saveVideoData } from "../services";
+import { saveVideoData } from "../api";
 
 export function useSaveVideoData() {
   const queryClient = useQueryClient();
