--- apps/client/src/features/video/services/index.ts
@@ -1,1 +1,0 @@
-export * from "./video.service";
